+++ modules/app-browser/src/App.jsx
@@ -1,9 +1,18 @@
-import React from "react";
+import React, { Component } from "react";
+import ClipList from "./ClipList";
 
-class App extends React.Component {
-  render() {
-    return <div>Hi</div>;
-  }
-}
+const SetMeta = ({ meta }) => (
+  <div className="meta">
+    <h1>{meta.title}</h1>
+    <p>{meta.description}</p>
+  </div>
+);
+
+const App = ({ set }) => (
+  <div className="Set">
+    <SetMeta meta={set.meta} />
+    <ClipList set={set} />
+  </div>
+);
 
 export default App;