+++ modules/audioset/src/index.js
@@ -1,2 +1,1 @@
-export { default as loadSetHoc } from "./load-set-hoc";
-export { default as wsConnectHoc } from "./ws-connect-hoc";
+export function loadSet(url) {}